Renamed Network to safeFetch
