+++ src/helpers/redditFeedHelpers.ts
@@ -3,7 +3,7 @@ import * as urlUtils from './urlUtils';
 import { Feed } from '../models/Feed';
 import { Debug } from '../Debug';
 import { fetchFaviconUrl } from './favicon';
-import { safeFetch } from '../Network';
+import { safeFetch } from './safeFetch';
 import { asyncTryExpr } from './tryExpr';
 import { isError } from 'util';
