+++ src/swarm/Swarm.ts
@@ -1,6 +1,6 @@
 
 import { Debug } from '../Debug';
-import { safeFetch } from '../Network';
+import { safeFetch } from '../helpers/safeFetch';
 
 export const defaultGateway = 'https://swarm.felfele.com';
 export const defaultPublicGateway = 'https://swarm-gateways.net';
